requirements.txt
ultralytics
opencv-python
numpy
flask
pyserial
requests